Move recordSubforumView into server directory
